First flying experiment published (build_20201105)
A lot of change in the scene and script to adjust into flying mode(switch from mainSpatialTask_ver03.cs to mainSpatialTask_ver4.cs). Change in Environment and scene 1)the glass fence was removed because I want to avoid the situation where subject bump into the fence and stuck in the corner between the the grass and fence (at the end of the curve). And subjects don't have to worry about falling into the water when they fly. However, I wonder whether I should have remained the fence to make the boundary information equivalent to the previous driving condition.. 2)I also adjusted the camera height of the character (previously camera was always 1 unit above the surface) so that camera rotation is pivoted at the center of the camera (an exception is the short demo video for straight line distance which shows arrow.  I think the rotation was pivoted below the camera as in the driving example.. but I didn't feel the rotation strange.) 3) I gave a slight vertical offset to picture cube object from the ground because with the previous offset bottom of the picture cube was shown from the opposite side of the surface and I don't want flying participant to see the picture cube pentrating the floor..(The more elegant way would be to give more thickness to the curved surface) 4)I removed the guiding arrow on the surface in the current setting. because in the flying condition, projecting guiding arrow on the surface doesn't make much sense. Due to the lack of guiding arrow plus generally more difficult control of the flying movement, I expect it would take longer to find object and complete the experiment in the current version. I also worried about subjects not finding the traffic cone or the picture cube, so I added the cone and picture cube in the instruction. I also added screenshots of Distance estimation tasks in the instruction set. I tried to do the most of works "for participants" to make the experiment as clear and intuitive as possible. 4)Rotation is pitch rotation along the world axis plus yaw rotation along the body axis. This creates unusual dizzying head rotation.. I spent a lot of time on how to implement 3D rotation using keyboards, what constraints should be allowed, but didn't come up with the one perfect solution. Let's see the result. 5)I also slightly adjusted the wording of egocentric/allocentric strategy in the debrief question.